add funtions to main

+++ ItemManager.cpp
@@ -137,10 +137,9 @@ string ItemManager::getItemFile()
 	return this->itemFile;
 }
 
-//Setter
-
-
-void ItemManager::setItemList(vector<Item*> itemList) {
+// Setter
+void ItemManager::setItemList(vector<Item *> itemList)
+{
 	this->itemList = itemList;
 }
 void ItemManager::setNoOfItem(int noOfItem)
@@ -176,7 +175,8 @@ bool ItemManager::addItem()
 	// generate item id - code part (3 digits)
 	ostringstream streamID;
 	streamID << "I" << setw(3) << setfill('0') << this->noOfItem;
-	string id = streamID.str();
+	int code = stoi((this->itemList.back())->getId().substr(1, 3));
+	string id = code.str();
 
 	id.append("-");
 
@@ -376,6 +376,13 @@ bool ItemManager::updateItem(string id)
 		}
 	}
 
+	string newId = item->getId();
+	string newTitle = item->getTitle();
+	string newLoanType = item->getLoanType();
+	string newRentalType = newValue;
+	int newNoOfCopy = item->getNoOfCopy();
+	double newFee = item->getFee();
+
 	switch (choice)
 	{
 	case '1':
@@ -384,23 +391,16 @@ bool ItemManager::updateItem(string id)
 		cin.ignore();
 		item->setTitle(newValue);
 		return true;
-	case '2';
+	case '2':
 		cout << "\nEnter the new loan type: ";
 		cin >> newValue;
 		cin.ignore();
 		item->setLoanType(newValue);
 		return true;
-		case "3":
+	case '3':
 		cout << "\nEnter the new rental type: ";
 		cin >> newValue;
 		cin.ignore();
-
-		string newId = item->getId();
-		string newTitle = item->getTitle();
-		string newLoanType = item->getLoanType();
-		string newRentalType = newValue;
-		int newNoOfCopy = item->getNoOfCopy();
-		double newFee = item->getFee();
 
 		switch (newValue)
 		{
@@ -412,6 +412,13 @@ bool ItemManager::updateItem(string id)
 			return true;
 		case "Record":
 			string newGenre = item->getGenre();
+			if (!newGenre || newGenre.size() == 0)
+			{
+				cout << "\nEnter the new genre: ";
+				cin >> newGenre;
+				cin.ignore();
+				break;
+			}
 			deleteItem(itemList, newId);
 			Item *newItem = new Record(newId, newTitle, newRentalType, newLoanType, newNoOfCopy, newFee, newGenre);
 			this->itemList.push_back(newItem);
@@ -452,66 +459,22 @@ bool ItemManager::updateItem(string id)
 	return false;
 }
 
-bool ItemManager::deleteItem(vector<Item *> &list, string id)
+bool ItemManager::deleteItem(string id)
 {
-	T search(const vector<T> &list, const string &id)
+	for (Item *item : this->itemList)
 	{
-		vector<T> tempVector;
-		for (auto i : list)
+		if (item->getId() == id)
 		{
-			if (i->getID() == id)
-			{
-				tempVector.push_back(i);
-			}
-		}
-		size_t size = tempVector.size();
-		if (tempVector.empty())
-		{
-			return nullptr;
-		}
-		else if (size == 1)
-		{
-			return tempVector.front();
+			this->itemList.erase(find(this->itemList.begin(), this->itemList.end(), item));
+			delete item;
+			return true;
 		}
 		else
 		{
-			cout << size << " matching IDs found! Please choose one ID: " << endl;
-
-			for (size_t i = 0; i < size; i++)
-			{
-				cout << i + 1 << ". " << tempVector[i]->getDescription() << endl;
-			}
-			size_t choice = -1;
-			while (true)
-			{
-				cout << "Enter your choice: ";
-				cin >> choice;
-				cin.ignore();
-				if (choice > 0 && choice <= size)
-				{
-					return tempVector[choice - 1];
-				}
-				else
-				{
-					cerr << "Invalid input! Please only choose between 1 and " << size << "!" << endl;
-				}
-			}
+			cerr << "No matching ID found!" << endl;
+			return false;
 		}
 	}
-
-	Item *deletedItem = search(list, idOrName);
-	if (deletedItem == nullptr)
-	{
-		cerr << "No matching ID found!" << endl;
-		return false;
-	}
-	else
-	{
-		list.erase(find(list.begin(), list.end(), deletedItem));
-		delete deletedItem;
-		return true;
-	}
-	return false;
 }
 
 void ItemManager::displaySortedItemTitle() {
@@ -520,12 +483,12 @@ void ItemManager::displaySortedItemTitle() {
 		for (int j = i+1; j < this->getItemList().size(); j++ ){
 			if (this->getItemList()[i]->getTitle()>this->getItemList()[j]->getTitle()){
 				string temp = this->getItemList()[i]->getTitle();
+
 				this->getItemList()[i]->getTitle() = this->getItemList()[j]->getTitle();
 				this->getItemList()[j]->getTitle() = temp;
 			}
 		}	
 	}
-	cout << "The item list sorted by Title is: ";
 	for (Item* item : this->getItemList()) {
 		cout << item->getTitle();
 	}
@@ -542,9 +505,8 @@ void ItemManager::displaySortedItemID()
 				this->getItemList()[i]->getId() = this->getItemList()[j]->getId();
 				this->getItemList()[j]->getId() = temp;
 			}
-		}	
+		}	orted
 	}
-	cout << "The item list sorted by ID is: ";
 	for (Item* item : this->getItemList()) {
 		cout << item->getId();
 	}
